+++ storage/src/test/java/tech/pegasys/teku/storage/server/kvstore/AbstractDatabaseTest.java
@@ -612,7 +612,7 @@ public abstract class AbstractDatabaseTest {
     recentChainData
         .getUpdatableForkChoiceStrategy()
         .orElseThrow()
-        .onExecutionPayloadResult(transitionBlock.getRoot(), PayloadStatus.VALID);
+        .onExecutionPayloadResult(transitionBlock.getRoot(), PayloadStatus.VALID, true);
 
     // Then finalize
     final Checkpoint finalizedCheckpoint =